# Add type enc/dec cache