--- a/.travis.yml
+++ b/.travis.yml
@@ -14,3 +14,6 @@ env:
   - GO111MODULE=on
 
 go_import_path: github.com/vmihailenco/msgpack
+
+before_install:
+  - curl -sfL https://install.goreleaser.com/github.com/golangci/golangci-lint.sh | sh -s -- -b $(go env GOPATH)/bin v1.17.1