--- a/Makefile
+++ b/Makefile
@@ -3,6 +3,4 @@ all:
 	go test ./... -short -race
 	go test ./... -run=NONE -bench=. -benchmem
 	env GOOS=linux GOARCH=386 go test ./...
-	go vet
-	go get github.com/gordonklaus/ineffassign
-	ineffassign .
+	golangci-lint run