--- a/decode_value.go
+++ b/decode_value.go
@@ -11,6 +11,7 @@ var stringType = reflect.TypeOf((*string)(nil)).Elem()
 
 var valueDecoders []decoderFunc
 
+//nolint:gochecknoinits
 func init() {
 	valueDecoders = []decoderFunc{
 		reflect.Bool:          decodeBoolValue,
@@ -83,12 +84,10 @@ func _getDecoder(typ reflect.Type) decoderFunc {
 		return ptrDecoderFunc(typ)
 	case reflect.Slice:
 		elem := typ.Elem()
-		switch elem.Kind() {
-		case reflect.Uint8:
+		if elem.Kind() == reflect.Uint8 {
 			return decodeBytesValue
 		}
-		switch elem {
-		case stringType:
+		if elem == stringType {
 			return decodeStringSliceValue
 		}
 	case reflect.Array: