--- a/encode_slice.go
+++ b/encode_slice.go
@@ -6,6 +6,8 @@ import (
 	"github.com/vmihailenco/msgpack/codes"
 )
 
+var sliceStringType = reflect.TypeOf(([]string)(nil))
+
 func encodeStringValue(e *Encoder, v reflect.Value) error {
 	return e.EncodeString(v.String())
 }
@@ -86,6 +88,11 @@ func (e *Encoder) EncodeArrayLen(l int) error {
 		return e.write2(codes.Array16, uint16(l))
 	}
 	return e.write4(codes.Array32, uint32(l))
+}
+
+func encodeStringSliceValue(e *Encoder, v reflect.Value) error {
+	ss := v.Convert(sliceStringType).Interface().([]string)
+	return e.encodeStringSlice(ss)
 }
 
 func (e *Encoder) encodeStringSlice(s []string) error {