```
benchmark                               old ns/op     new ns/op     delta
BenchmarkDiscard-4                      38.1          38.4          +0.79%
BenchmarkBool-4                         25.2          25.2          +0.00%
BenchmarkInt0-4                         76.4          76.1          -0.39%
BenchmarkInt1-4                         76.0          76.6          +0.79%
BenchmarkInt2-4                         76.1          76.0          -0.13%
BenchmarkInt4-4                         76.1          76.2          +0.13%
BenchmarkInt8-4                         75.7          75.7          +0.00%
BenchmarkInt32-4                        175           121           -30.86%
BenchmarkTime-4                         94.3          93.4          -0.95%
BenchmarkDuration-4                     74.9          75.6          +0.93%
BenchmarkByteSlice-4                    160           160           +0.00%
BenchmarkByteArray-4                    597           479           -19.77%
BenchmarkByteArrayPtr-4                 625           436           -30.24%
BenchmarkMapStringString-4              638           577           -9.56%
BenchmarkMapStringStringPtr-4           948           703           -25.84%
BenchmarkMapStringInterface-4           796           688           -13.57%
BenchmarkMapIntInt-4                    2224          1802          -18.97%
BenchmarkStringSlice-4                  557           423           -24.06%
BenchmarkStringSlicePtr-4               831           551           -33.69%
BenchmarkStructVmihailencoMsgpack-4     6484          5860          -9.62%
BenchmarkStructMarshal-4                3329          2693          -19.10%
BenchmarkStructUnmarshal-4              2315          2207          -4.67%
BenchmarkStructManual-4                 4462          3888          -12.86%
BenchmarkStructUnmarshalPartially-4     2186          2150          -1.65%
BenchmarkQuery-4                        3208          3223          +0.47%

benchmark                               old allocs     new allocs     delta
BenchmarkDiscard-4                      0              0              +0.00%
BenchmarkBool-4                         0              0              +0.00%
BenchmarkInt0-4                         0              0              +0.00%
BenchmarkInt1-4                         0              0              +0.00%
BenchmarkInt2-4                         0              0              +0.00%
BenchmarkInt4-4                         0              0              +0.00%
BenchmarkInt8-4                         0              0              +0.00%
BenchmarkInt32-4                        0              0              +0.00%
BenchmarkTime-4                         0              0              +0.00%
BenchmarkDuration-4                     0              0              +0.00%
BenchmarkByteSlice-4                    0              0              +0.00%
BenchmarkByteArray-4                    2              2              +0.00%
BenchmarkByteArrayPtr-4                 3              2              -33.33%
BenchmarkMapStringString-4              4              4              +0.00%
BenchmarkMapStringStringPtr-4           5              4              -20.00%
BenchmarkMapStringInterface-4           6              6              +0.00%
BenchmarkMapIntInt-4                    10             10             +0.00%
BenchmarkStringSlice-4                  2              2              +0.00%
BenchmarkStringSlicePtr-4               3              2              -33.33%
BenchmarkStructVmihailencoMsgpack-4     28             27             -3.57%
BenchmarkStructMarshal-4                10             9              -10.00%
BenchmarkStructUnmarshal-4              18             18             +0.00%
BenchmarkStructManual-4                 22             22             +0.00%
BenchmarkStructUnmarshalPartially-4     11             11             +0.00%
BenchmarkQuery-4                        3              3              +0.00%

benchmark                               old bytes     new bytes     delta
BenchmarkDiscard-4                      0             0             +0.00%
BenchmarkBool-4                         0             0             +0.00%
BenchmarkInt0-4                         0             0             +0.00%
BenchmarkInt1-4                         0             0             +0.00%
BenchmarkInt2-4                         0             0             +0.00%
BenchmarkInt4-4                         0             0             +0.00%
BenchmarkInt8-4                         0             0             +0.00%
BenchmarkInt32-4                        0             0             +0.00%
BenchmarkTime-4                         0             0             +0.00%
BenchmarkDuration-4                     0             0             +0.00%
BenchmarkByteSlice-4                    0             0             +0.00%
BenchmarkByteArray-4                    64            64            +0.00%
BenchmarkByteArrayPtr-4                 80            64            -20.00%
BenchmarkMapStringString-4              16            16            +0.00%
BenchmarkMapStringStringPtr-4           32            16            -50.00%
BenchmarkMapStringInterface-4           48            48            +0.00%
BenchmarkMapIntInt-4                    208           208           +0.00%
BenchmarkStringSlice-4                  10            10            +0.00%
BenchmarkStringSlicePtr-4               26            10            -61.54%
BenchmarkStructVmihailencoMsgpack-4     2240          2224          -0.71%
BenchmarkStructMarshal-4                1856          1840          -0.86%
BenchmarkStructUnmarshal-4              384           384           +0.00%
BenchmarkStructManual-4                 1808          1808          +0.00%
BenchmarkStructUnmarshalPartially-4     1312          1312          +0.00%
BenchmarkQuery-4                        72            72            +0.00%
```